+++ eosio_types/signature_type.mjs
@@ -3,15 +3,13 @@ import { GraphQLScalarType } from "graphql";
 const signature_type = new GraphQLScalarType({
   description: `\`Signature type\`
 
-  EOS K1 signature supported.`,
+  Antelope based signature, K1, R1 WA.`,
   name: "signature",
   parseValue(signature) {
     if (signature == "") return "";
     if (typeof signature !== "string")
       throw new TypeError("Expected signature to be string");
 
-    if (signature.slice(0, 7) != "SIG_K1_")
-      throw new TypeError("Signature prefix is should be SIG_K1_.");
     return signature;
   }
 });